There was a bug in my version of ContractData.js
```
Error: Invalid number of arguments to Solidity function
```
It was caused because I forgot to apply spread syntax in line 18 of
ContractData.js
```javascript
dataKey: this.method.cacheCall(...this.props.methodArgs)
```
Now the ContractData component can update whenever you change its props
without having to re-mount the component. This is a small upgrade to a
contrubution by @parkinsonj, all the credits for the re-writing of
drizzle-react-components for React Context API v16.3+ go to him!